+++ crates/pretty_text/src/lib.rs
@@ -58,7 +58,7 @@
 //!
 //!     // Parse rich text and use custom style.
 //!     commands.spawn((
-//!         pretty!("I am [1]<0.8>*sniff*[1]<1.2> very `pretty`[my_style]![3]<1>"),
+//!         pretty!("I am [1]<0.8>*sniff*[1]<1.2> very [pretty](my_style)![3]<1>"),
 //!         Transform::from_xyz(0.0, -200.0, 0.0),
 //!     ));
 //! }